src/jobboard_app/core/business_logic/services/common.py
from __future__ import annotations

import sys
import uuid
from io import BytesIO
from typing import TYPE_CHECKING

from django.core.files.uploadedfile import InMemoryUploadedFile
from PIL import Image

if TYPE_CHECKING:
    from django.core.files import File


def replace_file_name_to_uuid(file: File) -> File:
    file_extension = file.name.split(".")[-1]
    file_name = str(uuid.uuid4())
    file.name = file_name + "." + file_extension
    return file


def change_file_size(file: InMemoryUploadedFile) -> InMemoryUploadedFile:
    format = file.content_type.split("/")[-1].upper()
    print(format)
    output = BytesIO()
    with Image.open(file) as image:
        image.thumbnail(size=(250, 200))
        image.save(output, format=format, quality=100)

    return InMemoryUploadedFile(
        file=output,
        field_name=file.field_name,
        name=file.name,
        content_type=file.content_type,
        size=sys.getsizeof(output),
        charset=file.charset,
    )
