pyproject.toml
[tool.poetry]
name = "job-board-project"
version = "0.1.0"
description = ""
authors = ["KaterinaPy35 <yakushka1308@mail.ru>"]
readme = "README.md"
packages = [{include = "job_board_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.2"
python-dotenv = "1.0.0"
pillow = "10.0.0"
dacite = "1.8.1"
psycopg2-binary = "2.9.6"


[tool.poetry.group.dev.dependencies]
pylint = "2.17.4"
deptry = "0.11.0"
pre-commit = "3.3.3"

[tool.mypy]
mypy_path = "src/jobboard_app"
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true
no_implicit_optional = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = true

[tool.black]
skip-string-normalization = false
line-length = 120

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
